// src/enums/interfaces/IWaiterCommissionReport.ts
export interface IWaiterCommissionReport {
  waiter_id: number
  waiter_name: string
  waiter_email: string
  total_items_sold: number
  others_items: number
  food_items: number
  beverages_items: number
  total_revenue: number
  others_commission: number
  food_commission: number
  beverages_commission: number
  total_commission: number
}